fix test and merge channels.go and users_private_channels.go
